requirements.txt
numpy==1.26.4
scipy==1.11.4
pydicom==2.4.4
SimpleITK==2.3.1
scikit-image==0.22.0
torch==2.2.2
monai==1.3.2
meshio==5.3.5
pyvista==0.43.10
vtk==9.3.0
tetgen==0.6.4
gmsh==4.13.1
pygmsh==7.1.17
matplotlib==3.8.4
reportlab==4.2.2
streamlit==1.35.0
PyYAML==6.0.1
tqdm==4.66.4
pandas==2.2.2
crewai==0.41.1
pytest==8.2.2

